+++ frontend/openchat-agent/src/services/notifications/candid/types.d.ts
@@ -82,6 +82,7 @@ export interface ChannelMembershipUpdates {
 export interface ChannelMessageNotification {
   'channel_id' : ChannelId,
   'community_id' : CommunityId,
+  'message_sub_type' : [] | [string],
   'image_url' : [] | [string],
   'sender_display_name' : [] | [string],
   'sender' : UserId,
@@ -392,6 +393,7 @@ export interface DirectChatSummaryUpdates {
   'latest_message' : [] | [MessageEventWrapper],
 }
 export interface DirectMessageNotification {
+  'message_sub_type' : [] | [string],
   'image_url' : [] | [string],
   'sender_display_name' : [] | [string],
   'sender_avatar_id' : [] | [bigint],
@@ -627,6 +629,7 @@ export interface GroupMatch {
   'member_count' : number,
 }
 export interface GroupMessageNotification {
+  'message_sub_type' : [] | [string],
   'image_url' : [] | [string],
   'group_avatar_id' : [] | [bigint],
   'sender_display_name' : [] | [string],