+++ frontend/openchat-shared/src/domain/notifications/index.ts
@@ -38,6 +38,7 @@ export type ChannelNotification = {
     communityName: string;
     channelName: string;
     messageType: string;
+    messageSubType: string | undefined;
     messageText: string | undefined;
     imageUrl: string | undefined;
     communityAvatarId: bigint | undefined;
@@ -54,6 +55,7 @@ export type DirectNotification = {
     senderName: string;
     senderDisplayName: string | undefined;
     messageType: string;
+    messageSubType: string | undefined;
     messageText: string | undefined;
     imageUrl: string | undefined;
     senderAvatarId: bigint | undefined;
@@ -72,6 +74,7 @@ export type GroupNotification = {
     chatId: GroupChatIdentifier;
     groupName: string;
     messageType: string;
+    messageSubType: string | undefined;
     messageText: string | undefined;
     imageUrl: string | undefined;
     groupAvatarId: bigint | undefined;