+++ backend/canisters/community/impl/src/updates/claim_prize.rs
@@ -141,7 +141,7 @@ fn commit(args: Args, winner: UserId, transaction: CompletedCryptoTransaction, s
                 sender: OPENCHAT_BOT_USER_ID,
                 sender_name: OPENCHAT_BOT_USERNAME.to_string(),
                 sender_display_name: None,
-                message_type: content.message_type().to_string(),
+                message_type: content.message_type(),
                 message_text: content.notification_text(&[], &[]),
                 image_url: content.notification_image_url(),
                 community_avatar_id: state.data.avatar.as_ref().map(|d| d.id),