+++ backend/canisters/community/impl/src/updates/send_message.rs
@@ -116,7 +116,7 @@ fn send_message_impl(args: Args, state: &mut RuntimeState) -> Response {
                     sender: user_id,
                     sender_name: args.sender_name,
                     sender_display_name: member.display_name().value.clone().or(args.sender_display_name),
-                    message_type: content.message_type().to_string(),
+                    message_type: content.message_type(),
                     message_text: content.notification_text(
                         &args.mentioned,
                         &user_groups_mentioned