+++ backend/libraries/types/src/message_content.rs
@@ -108,8 +108,8 @@ impl MessageContent {
         references
     }
 
-    pub fn message_type(&self) -> &'static str {
-        match self {
+    pub fn message_type(&self) -> String {
+        let message_type = match self {
             MessageContent::Text(_) => "Text",
             MessageContent::Image(_) => "Image",
             MessageContent::Video(_) => "Video",
@@ -125,8 +125,10 @@ impl MessageContent {
             MessageContent::MessageReminderCreated(_) => "MessageReminderCreated",
             MessageContent::MessageReminder(_) => "MessageReminder",
             MessageContent::ReportedMessage(_) => "ReportedMessage",
-            MessageContent::Custom(_) => "Custom",
-        }
+            MessageContent::Custom(c) => &c.kind,
+        };
+
+        message_type.to_string()
     }
 
     pub fn text(&self) -> Option<&str> {